push-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
jinja2
aio-pika
httpx
aioredis
pyfcm
python-dotenv
